updating for dhamma

+++ credit/physics_constants.py
@@ -5,34 +5,34 @@ Reference:
 
 Harrop, B.E., Pritchard, M.S., Parishani, H., Gettelman, A., Hagos, S., Lauritzen, P.H., Leung, L.R., Lu, J., Pressel, K.G. and Sakaguchi, K., 2022. Conservation of dry air, water, and energy in CAM and its potential impact on tropical rainfall. Journal of Climate, 35(9), pp.2895-2917.
 """
-
+import numpy as np
 # Earth's radius
-RAD_EARTH = 6371000  # m
+RAD_EARTH = np.float64(6371000 ) # m
 
 # ideal gas constant of water vapor
-RVGAS = 461.5  # J/kg/K
+RVGAS = np.float64(461.5)  # J/kg/K
 
 # ideal gas constant of dry air
-RDGAS = 287.05  # J/kg/K
+RDGAS = np.float64(287.05)  # J/kg/K
 
 # gravity
-GRAVITY = 9.80665  # m/s^2
+GRAVITY = np.float64(9.80665)  # m/s^2
 
 # density of water
-RHO_WATER = 1000.0  # kg/m^3
+RHO_WATER = np.float64(1000.0)  # kg/m^3
 
 # ========================================================= #
 # latent heat caused by the phase change of water (0 deg C)
-LH_WATER = 2.501e6  # J/kg
-LH_ICE = 333700  # J/kg
+LH_WATER = np.float64(2.501e6)  # J/kg
+LH_ICE = np.float64(333700)  # J/kg
 
 # ========================================================= #
 # heat capacity on constant pressure
 # dry air
-CP_DRY = 1004.64  # J/kg K
+CP_DRY = np.float64(1004.64)  # J/kg K
 # water vapor
-CP_VAPOR = 1810.0  # J/kg K
+CP_VAPOR = np.float64(1810.0)  # J/kg K
 # liquid
-CP_LIQUID = 4188.0  # J/kg K
+CP_LIQUID = np.float64(4188.0)  # J/kg K
 # ice
-CP_ICE = 2117.27  # J/kg K
+CP_ICE = np.float64(2117.27)  # J/kg K
